# frontend/lib/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Route 53 Clone API",
    "description": "\nBackend of the Route 53 console clone. Hosted zones and record sets follow Route 53's\nrules: every zone owns an apex NS and SOA record, CNAMEs cannot share a name with other\nrecords, and change batches are atomic.\n\nSign in with `POST /api/v1/auth/login`; the session travels in an httpOnly cookie.\nErrors always have the shape `{code, message, details}`.\n",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/hostedzones": {
      "get": {
        "tags": [
          "Hosted zones"
        ],
        "summary": "List hosted zones",
        "operationId": "list_hosted_zones",
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 255
                },
                {
                  "type": "null"
                }
              ],
              "description": "Free-text search across the main columns",
              "title": "Search"
            },
            "description": "Free-text search across the main columns"
          },
          {
            "name": "type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ZoneType"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Only zones of this type",
              "title": "Type"
            },
            "description": "Only zones of this type"
          },
          {
            "name": "filter",
            "in": "query",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              },
              "description": "Property filter as `field:operator:value`, repeatable. Operators: eq, ne, contains, not_contains, starts_with, not_starts_with, gt, gte, lt, lte.",
              "examples": [
                "name:contains:shop"
              ],
              "default": [],
              "title": "Filter"
            },
            "description": "Property filter as `field:operator:value`, repeatable. Operators: eq, ne, contains, not_contains, starts_with, not_starts_with, gt, gte, lt, lte."
          },
          {
            "name": "filter_mode",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "and",
                "or"
              ],
              "type": "string",
              "description": "Whether all filters must match or any of them",
              "default": "and",
              "title": "Filter Mode"
            },
            "description": "Whether all filters must match or any of them"
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "default (Route 53 order), name, type, description, id, created_by, record_count or created_at",
              "default": "default",
              "title": "Sort"
            },
            "description": "default (Route 53 order), name, type, description, id, created_by, record_count or created_at"
          },
          {
            "name": "order",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "asc",
                "desc"
              ],
              "type": "string",
              "description": "Sort direction",
              "default": "asc",
              "title": "Order"
            },
            "description": "Sort direction"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "1-based page number",
              "default": 1,
              "title": "Page"
            },
            "description": "1-based page number"
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "description": "Items per page",
              "default": 50,
              "title": "Page Size"
            },
            "description": "Items per page"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_HostedZoneSummary_"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in or the session expired",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "400": {
            "description": "The request violates a Route 53 rule",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "The request body or parameters are malformed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Hosted zones"
        ],
        "summary": "Create a hosted zone",
        "description": "Also creates the apex NS and SOA record sets, as Route 53 does.",
        "operationId": "create_hosted_zone",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HostedZoneCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HostedZoneDetail"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in or the session expired",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "400": {
            "description": "The request violates a Route 53 rule",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "The request body or parameters are malformed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/hostedzones/{zone_id}": {
      "get": {
        "tags": [
          "Hosted zones"
        ],
        "summary": "Get a hosted zone",
        "operationId": "get_hosted_zone",
        "parameters": [
          {
            "name": "zone_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Hosted zone ID",
              "examples": [
                "Z0123456789ABCDEFGHIJ"
              ],
              "title": "Zone Id"
            },
            "description": "Hosted zone ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HostedZoneDetail"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in or the session expired",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "The hosted zone does not exist",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Hosted zones"
        ],
        "summary": "Edit a hosted zone",
        "description": "The description, the tags and, for a private zone, the VPC associations can be changed after a zone is created. Omitted fields keep their value; the edit is applied as a whole or not at all.",
        "operationId": "update_hosted_zone",
        "parameters": [
          {
            "name": "zone_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Hosted zone ID",
              "examples": [
                "Z0123456789ABCDEFGHIJ"
              ],
              "title": "Zone Id"
            },
            "description": "Hosted zone ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HostedZoneUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HostedZoneDetail"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in or the session expired",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "The hosted zone does not exist",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "400": {
            "description": "The request violates a Route 53 rule",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "The request body or parameters are malformed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Hosted zones"
        ],
        "summary": "Delete a hosted zone",
        "description": "Fails with `HostedZoneNotEmpty` while the zone has records besides the apex NS and SOA.",
        "operationId": "delete_hosted_zone",
        "parameters": [
          {
            "name": "zone_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Hosted zone ID",
              "examples": [
                "Z0123456789ABCDEFGHIJ"
              ],
              "title": "Zone Id"
            },
            "description": "Hosted zone ID"
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "description": "Not signed in or the session expired",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "The hosted zone does not exist",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "The change conflicts with existing data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/hostedzones/{zone_id}/tags": {
      "get": {
        "tags": [
          "Hosted zones",
          "Tags"
        ],
        "summary": "List a zone's tags",
        "operationId": "list_tags",
        "parameters": [
          {
            "name": "zone_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Hosted zone ID",
              "examples": [
                "Z0123456789ABCDEFGHIJ"
              ],
              "title": "Zone Id"
            },
            "description": "Hosted zone ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TagList"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in or the session expired",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "The hosted zone does not exist",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Hosted zones",
          "Tags"
        ],
        "summary": "Replace a zone's tags",
        "operationId": "replace_tags",
        "parameters": [
          {
            "name": "zone_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Hosted zone ID",
              "examples": [
                "Z0123456789ABCDEFGHIJ"
              ],
              "title": "Zone Id"
            },
            "description": "Hosted zone ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TagsUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TagList"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in or the session expired",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "The hosted zone does not exist",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "400": {
            "description": "The request violates a Route 53 rule",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "The request body or parameters are malformed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorDetail": {
        "properties": {
          "field": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Field",
            "description": "Request field the problem relates to"
          },
          "index": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Index",
            "description": "Position of the failing batch change"
          },
          "line": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Line",
            "description": "Zone file line the problem relates to"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "ErrorDetail"
      },
      "ErrorResponse": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "examples": [
              "NoSuchHostedZone"
            ]
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "details": {
            "items": {
              "$ref": "#/components/schemas/ErrorDetail"
            },
            "type": "array",
            "title": "Details",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "HostedZoneCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 1024,
            "minLength": 1,
            "title": "Name",
            "examples": [
              "example.com"
            ]
          },
          "description": {
            "type": "string",
            "maxLength": 256,
            "title": "Description",
            "default": ""
          },
          "type": {
            "$ref": "#/components/schemas/ZoneType",
            "default": "public"
          },
          "vpcs": {
            "items": {
              "$ref": "#/components/schemas/VpcAssociation"
            },
            "type": "array",
            "title": "Vpcs",
            "default": []
          },
          "tags": {
            "items": {
              "$ref": "#/components/schemas/Tag"
            },
            "type": "array",
            "title": "Tags",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "HostedZoneCreate"
      },
      "HostedZoneDetail": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "examples": [
              "Z0123456789ABCDEFGHIJ"
            ]
          },
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Canonical zone name with trailing dot",
            "examples": [
              "example.com."
            ]
          },
          "type": {
            "$ref": "#/components/schemas/ZoneType"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "created_by": {
            "type": "string",
            "title": "Created By"
          },
          "record_count": {
            "type": "integer",
            "title": "Record Count"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "caller_reference": {
            "type": "string",
            "title": "Caller Reference"
          },
          "name_servers": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Name Servers"
          },
          "vpcs": {
            "items": {
              "$ref": "#/components/schemas/VpcAssociation"
            },
            "type": "array",
            "title": "Vpcs"
          },
          "tags": {
            "items": {
              "$ref": "#/components/schemas/Tag"
            },
            "type": "array",
            "title": "Tags"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "type",
          "description",
          "created_by",
          "record_count",
          "created_at",
          "updated_at",
          "caller_reference",
          "name_servers",
          "vpcs",
          "tags"
        ],
        "title": "HostedZoneDetail"
      },
      "HostedZoneSummary": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "examples": [
              "Z0123456789ABCDEFGHIJ"
            ]
          },
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Canonical zone name with trailing dot",
            "examples": [
              "example.com."
            ]
          },
          "type": {
            "$ref": "#/components/schemas/ZoneType"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "created_by": {
            "type": "string",
            "title": "Created By"
          },
          "record_count": {
            "type": "integer",
            "title": "Record Count"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "type",
          "description",
          "created_by",
          "record_count",
          "created_at",
          "updated_at"
        ],
        "title": "HostedZoneSummary"
      },
      "HostedZoneUpdate": {
        "properties": {
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 256
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "tags": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/Tag"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tags",
            "description": "Replaces every tag of the zone"
          },
          "vpcs": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/VpcAssociation"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vpcs",
            "description": "Replaces the VPC associations of a private zone"
          }
        },
        "type": "object",
        "title": "HostedZoneUpdate",
        "description": "Partial update; omitted fields keep their current value."
      },
...
      "Page_HostedZoneSummary_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/HostedZoneSummary"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total",
            "description": "Number of items matching the query across all pages"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          },
          "pages": {
            "type": "integer",
            "title": "Pages"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "page_size",
          "pages"
        ],
        "title": "Page[HostedZoneSummary]"
      },
...
      "Tag": {
        "properties": {
          "key": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Key"
          },
          "value": {
            "type": "string",
            "maxLength": 256,
            "title": "Value",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "key"
        ],
        "title": "Tag"
      },
      "TagList": {
        "properties": {
          "tags": {
            "items": {
              "$ref": "#/components/schemas/Tag"
            },
            "type": "array",
            "title": "Tags"
          }
        },
        "type": "object",
        "required": [
          "tags"
        ],
        "title": "TagList"
      },
      "TagsUpdate": {
        "properties": {
          "tags": {
            "items": {
              "$ref": "#/components/schemas/Tag"
            },
            "type": "array",
            "title": "Tags"
          }
        },
        "type": "object",
        "required": [
          "tags"
        ],
        "title": "TagsUpdate"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VpcAssociation": {
        "properties": {
          "vpc_id": {
            "type": "string",
            "maxLength": 32,
            "minLength": 1,
            "title": "Vpc Id",
            "examples": [
              "vpc-0a1b2c3d4e5f67890"
            ]
          },
          "region": {
            "type": "string",
            "maxLength": 32,
            "minLength": 1,
            "title": "Region",
            "examples": [
              "us-east-1"
            ]
          }
        },
        "type": "object",
        "required": [
          "vpc_id",
          "region"
        ],
        "title": "VpcAssociation"
      },
...
      "ZoneType": {
        "type": "string",
        "enum": [
          "public",
          "private"
        ],
        "title": "ZoneType"
      }
    }
  }
}